Separate article from liveblog controller
