+++ article/app/controllers/ArticleControllers.scala
@@ -16,4 +16,5 @@ trait ArticleControllers {
   lazy val bookSectionAgent: NewspaperBookSectionTagAgent = wire[NewspaperBookSectionTagAgent]
   lazy val publicationController = wire[PublicationController]
   lazy val articleController = wire[ArticleController]
+  lazy val liveBlogController = wire[LiveBlogController]
 }
