+++ article/app/renderers/RemoteRender.scala
@@ -8,36 +8,39 @@ import model.{Cached, ContentFields, PageWithStoryPackage}
 import play.api.libs.json.Json
 import play.api.libs.ws.WSClient
 import play.api.mvc.{RequestHeader, Result}
-import scala.concurrent.Future
+
+import scala.concurrent.{ExecutionContext, Future}
 import play.twirl.api.Html
+
 import scala.concurrent.duration._
 import common.RichRequestHeader
+import ExecutionContext.Implicits.global
 
 object RemoteRender {
 
-  // todo: inject WS
-
-  def remoteRenderArticle(ws:WSClient, payload: String): Future[String] = ws.url(Configuration.rendering.renderingEndpoint)
-    .withRequestTimeout(2000.millis)
-    .addHttpHeaders("Content-Type" -> "application/json")
-    .post(payload)
-    .map((response) =>
-      response.body
-    )
-
-  def remoteRender(ws:WSClient, path: String, model: PageWithStoryPackage)(implicit request: RequestHeader): Future[Result] = model match {
-
-    case article : ArticlePage =>
-      val contentFieldsJson = if (request.isGuui) List("contentFields" -> Json.toJson(ContentFields(article.article))) else List()
-      val jsonResponse = () => List(("html", views.html.fragments.articleBody(article))) ++ contentFieldsJson
-      val jsonPayload = JsonComponent.jsonFor(model, jsonResponse():_*)
-
-      remoteRenderArticle(ws, jsonPayload).map(s => {
-        Cached(article){ RevalidatableResult.Ok(Html(s)) }
-      })
-
-    case _ => throw new Exception("Remote render not supported for this content type")
-
-  }
+//  // todo: inject WS
+//
+//  def remoteRenderArticle(ws:WSClient, payload: String): Future[String] = ws.url(Configuration.rendering.renderingEndpoint)
+//    .withRequestTimeout(2000.millis)
+//    .addHttpHeaders("Content-Type" -> "application/json")
+//    .post(payload)
+//    .map((response) =>
+//      response.body
+//    )
+//
+//  def remoteRender(ws:WSClient, path: String, model: PageWithStoryPackage)(implicit request: RequestHeader): Future[Result] = model match {
+//
+//    case article : ArticlePage =>
+//      val contentFieldsJson = if (request.isGuui) List("contentFields" -> Json.toJson(ContentFields(article.article))) else List()
+//      val jsonResponse = () => List(("html", views.html.fragments.articleBody(article))) ++ contentFieldsJson
+//      val jsonPayload = JsonComponent.jsonFor(model, jsonResponse():_*)
+//
+//      remoteRenderArticle(ws, jsonPayload).map(s => {
+//        Cached(article){ RevalidatableResult.Ok(Html(s)) }
+//      })
+//
+//    case _ => throw new Exception("Remote render not supported for this content type")
+//
+//  }
 
 }
